requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
ca-certificates=2021.4.13=h06a4308_1
cachetools=4.2.1=pypi_0
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.5=pypi_0
cryptography=3.4.7=pypi_0
et_xmlfile=1.0.1=py_1001
future=0.18.2=pypi_0
jdcal=1.4.1=py_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
msoffcrypto-tool=4.11.0=pypi_0
ncurses=6.2=he6710b0_1
olefile=0.46=pypi_0
openpyxl=3.0.7=pyhd3eb1b0_0
openssl=1.1.1k=h27cfd23_0
pefile=2019.4.18=pypi_0
pip=21.0.1=py37h06a4308_0
py2exe=0.10.3.1=pypi_0
pycparser=2.20=pypi_0
pysimplegui=4.38.0=pyh44b312d_0
python=3.7.10=hdb3f193_0
python_abi=3.7=1_cp37m
readline=8.1=h27cfd23_0
setuptools=52.0.0=py37h06a4308_0
sqlite=3.35.4=hdfb4753_0
tk=8.6.10=hbc83047_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
